% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\docType{data}
\name{cb_gray}
\alias{cb_gray}
\title{Color palettes for plots}
\format{An object of class \code{character} of length 1.}
\usage{
cb_gray
}
\description{
Color palettes and plot colors I like. Color blind-friendly palettes taken
from the Cookbook for R (http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/#palettes-color-brewer).
The Matplotlib default color palette was taken from the documentation (https://matplotlib.org/users/dflt_style_changes.html#colors-in-default-property-cycle).
}
\examples{
ggplot(mtcars, aes(wt, mpg)) + geom_point(aes(color = as.factor(cyl))) + scale_color_manual(values = cbPalette)
}
\keyword{color}
\keyword{ggplot2}
\keyword{palette}
\keyword{theme}
