% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_mat_df.R
\name{convert_mat_df}
\alias{convert_mat_df}
\title{Convert connectome matrix data to long form tidy data}
\usage{
convert_mat_df(mat)
}
\arguments{
\item{mat}{Matrix data to convert}
}
\description{
Converts connectome matrix data type (e.g. from Matlab or Numpy) to
ggplot2-friendly format for use with geom_tile(). Assumes rownames and colnames
exist and are the ROI labels.
}
\examples{
foo <- convert_mat_df(matrix(data=rnorm(n=25), nrow=5, ncol=5))
}
\keyword{ggplot2}
\keyword{matrix}
\keyword{plot}
