% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fill_na}
\alias{fill_na}
\title{Combine list of vectors filling with NA's}
\usage{
fill_na(x)
}
\arguments{
\item{x}{a list of vectors.}
}
\value{
a data.frame with vectors as columns.
}
\description{
Combines a list of vectors, filling with NA's to reach the longest vector. Returns a data.frame
}
\examples{
# Combine vectors a, b, and c:
a <- rnorm(10)
b <- rnorm(5)
c <- rnorm(9)
fill_na(list(a,b,c))
}
