% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{get_all_markers}
\alias{get_all_markers}
\title{Manipulation with markers}
\usage{
get_all_markers(
  seu,
  filename = NULL,
  positive = T,
  min_n = 10,
  alpha = 0.05,
  logfold = 1
)
}
\arguments{
\item{seu}{Seurat object}

\item{filename}{When specified, save into a file}

\item{positive}{Only return positive markers (Default: True)}

\item{min_n}{Number of markers per identity (Default: 10)}

\item{alpha}{Threshhold for adjusted p-value (Default: 0.05)}

\item{logfold}{Minimum required logfold change (Default: 1)}
}
\description{
Manipulation with markers
}
