% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_pallet}
\alias{get_pallet}
\title{Get pretty color pallet}
\usage{
get_pallet(n = NULL)
}
\arguments{
\item{n}{Number of colors}
}
\value{
List of colors (maximum supported: 15). If higher number is requested
function returns and empty list
}
\description{
Get pretty color pallet
}
