% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagesc.R
\name{imagesc}
\alias{imagesc}
\title{Plot matrix using ggplot2}
\usage{
imagesc(df, scale_label = "", cmap_option = "viridis")
}
\arguments{
\item{df}{Matrix data}

\item{scale_label}{Label to be applied to the colorbar scale (blank by default)}

\item{cmap_option}{Choose colormap among viridis scales (default)}
}
\description{
Plot matrix analagous to Matlab's imagesc function.
}
\examples{
foo <- convert_mat_df(matrix(data=rnorm(n=25), nrow=5, ncol=5))
imagesc(foo)
}
\keyword{ggplot2}
\keyword{matrix}
