% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_installed_packages.R
\name{list_installed_packages}
\alias{list_installed_packages}
\title{List user-installed packages (apart from base installation)}
\usage{
list_installed_packages()
}
\description{
See title.
}
\examples{
list_installed_packages()
}
\keyword{installed}
\keyword{packages}
