% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{map_significance}
\alias{map_significance}
\title{Helper function for `plot_significant`}
\usage{
map_significance(pval, asterisk, pos_only)
}
\arguments{
\item{asterisk}{Replace p-values with asterisks (Default: False)}

\item{pos_only}{Boolean flag}
}
\description{
Helper function for `plot_significant`
}
\keyword{internal}
