% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{msub}
\alias{msub}
\title{Multiple substitution of values}
\usage{
msub(x, old, new)
}
\arguments{
\item{x}{a numerical vector, matrix or data frame.}

\item{old}{a vector of values to substitute.}

\item{new}{a vector of values that will substitute old values.}
}
\value{
a vector with new values.
}
\description{
Makes a pair-wise substitions of a vector of values by another vector of values of the same length.
}
\examples{
# Replace 1 for 2, 2 for 3 and 3 for 4.
msub(1:3,1:3,2:4)
}
