% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiggsave.R
\name{multiggsave}
\alias{multiggsave}
\title{Save ggplot in multiple file formats simultaneously.}
\usage{
multiggsave(
  filename,
  width = dev.size()[1],
  height = dev.size()[2],
  filetypes = c("pdf", "png")
)
}
\arguments{
\item{filename}{Name of file}

\item{width}{Width of plot (defaults to current graphics device width)}

\item{height}{Height of plot (defaults to current graphics device height)}

\item{filetypes}{Vector of file types to save.}
}
\description{
Save ggplot in multiple file format simultaneously. User can input the
specific file types. Defaults are PDF and PNG.
}
\examples{
ggplot(mtcars, aes(mpg, wt)) + geom_point()
multiggsave("multiggsave-test")
}
\keyword{ggplot2}
\keyword{ggsave}
\keyword{plot}
\keyword{save}
