% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_theme.R
\name{my_theme}
\alias{my_theme}
\title{Custom ggplot2 theme}
\usage{
my_theme()
}
\description{
Some simple tweaks to the default ggplot2 theme for exploratory data analysis
or informal presentations.
}
\examples{
ggplot(mtcars, aes(wt, mpg)) + geom_point() + my_theme()
}
\keyword{ggplot2}
\keyword{theme}
