% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{non_zero_median}
\alias{non_zero_median}
\title{Non-zero median of data frames, matrices or vectors}
\usage{
non_zero_median(x, by = NULL)
}
\arguments{
\item{x}{a numerical vector, matrix or data frame.}

\item{by}{a vector or list of grouping elements, each as long as the variables in the data frame \code{x}. The elements are coerced to factors before use.}
}
\value{
Non zero median of \code{x} or non zero median of groups of columns in \code{x}.
}
\description{
Calculate non zero median of a data frame, matrix or vector. If the parameter \code{by} is applied, median will be calculated per groups of columns, not rows.
}
\examples{
vector_test <- c(rep(0,10),1:100)
non_zero_median(vector_test)
matrix_test <- matrix(c(rep(0,10),1:100),nrow=10)
non_zero_median(matrix_test, by = c(1,2,3,1,2,3,1,2,3,1,2))
df_test <- as.data.frame(matrix_test)
non_zero_median(df_test, by = c(1,2,3,1,2,3,1,2,3,1,2))

}
