% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{nz_mean}
\alias{nz_mean}
\title{Non-zero mean of data frames, matrices or vectors}
\usage{
nz_mean(x, by = NULL)
}
\arguments{
\item{x}{a numerical vector, matrix or data frame.}

\item{by}{a vector or list of grouping elements, each as long as the variables in the data frame \code{x}. The elements are coerced to factors before use.}
}
\value{
Non zero mean of \code{x} or non zero mean of groups of columns in \code{x}.
}
\description{
Calculate non zero mean of a data frame, matrix or vector. If the parameter \code{by} is applied, mean will be calculated per groups of columns, not rows.
}
\examples{
vector_test <- c(rep(0,10),1:100)
nz_mean(vector_test)
matrix_test <- matrix(c(rep(0,10),1:100),nrow=10)
nz_mean(matrix_test, by = c(1,2,3,1,2,3,1,2,3,1,2))
df_test <- as.data.frame(matrix_test)
nz_mean(df_test, by = c(1,2,3,1,2,3,1,2,3,1,2))

}
