% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Visualize gene expression using boxplot}
\usage{
plot_boxplot(seu, features = NULL, slot = "data")
}
\arguments{
\item{seu}{Seurat object}

\item{features}{Genes of interest}

\item{slot}{Which matrix to use (Default: 'data')}
}
\description{
Visualize gene expression using boxplot
}
