% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_proportion_bar}
\alias{plot_proportion_bar}
\title{Visualize cellular proportion}
\usage{
plot_proportion_bar(seu, group.by = NULL)
}
\arguments{
\item{seu}{Seurat object}

\item{group.by}{Column name from metadata}
}
\description{
Visualize cellular proportion
}
