% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_violin_sig}
\alias{plot_violin_sig}
\title{Various plots}
\usage{
plot_violin_sig(seu, gene, conditions = NULL, asterisk = F)
}
\arguments{
\item{seu}{Seurat object}

\item{gene}{Gene of interest}

\item{conditions}{Specify which conditions to compare, otherwise use 
ones specified with `Idents()` (Default: NULL)}

\item{asterisk}{Replace p-values with asterisks (Default: False)}
}
\description{
Various plots
}
