% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_to_mm.R
\name{pt_to_mm}
\alias{pt_to_mm}
\title{Convert from pt to mm for ggplot2}
\usage{
pt_to_mm(x_pt)
}
\description{
ggplot2 uses mm by default for line units, this function takes the desired
thickness in pt and converts to mm.
}
\examples{
ggplot(mtcars, aes(wt, mpg)) + geom_line(size=pt_to_mm(0.75))
}
\keyword{ggplot2}
\keyword{mm}
\keyword{pt}
\keyword{unit}
