\name{require.r.config}
\alias{require.r.config}
\title{
require.r.config
}
\description{
set commonr options
}
\usage{
require.r.config(...)
}
\arguments{
  \item{\dots}{
  key=value pairs of config parameters that you would like to set
}
}
\details{
Supported arguments are:

basedir: The directory (relative from the location of the executing script) containing the modules
paths: A list with keys being the module names and values being the paths pointing to the respective modules
}
\author{
Stefan Seemayer
}

%% ~Make other sections like Warning with \section{Warning }{....} ~
