% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick-save.R
\name{save_1col}
\alias{save_1col}
\title{Save plots in 1-, 1.5-, or 2-column width for manuscripts}
\usage{
save_1col(title, path = ".", scale = 1)
}
\description{
Wraps ggsave to save plots in specified column widths for publication.
}
\examples{
p <- ggplot(mtcars, aes(mpg, wt)) + geom_point()
save_1col("mtcars-1col.png")
save_1.5col("mtcars-1.5col.png")
save_2col("mtcars-2col.png")
}
\keyword{ggplot2}
\keyword{ggsave}
\keyword{plot}
\keyword{save}
