% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_col.R
\name{save_col}
\alias{save_col}
\title{Save plots in 1-, 1.5-, or 2-column width for manuscripts}
\usage{
save_col(title, col_width = "1", path = ".", scale = 1)
}
\arguments{
\item{title}{Title of plot}

\item{col_width}{Column width (default = 1)}

\item{path}{Save path (default is getwd() value)}

\item{scale}{Scale of plot (default = 1)}
}
\description{
Wraps ggsave to save plots in specified column widths for publication.
}
\examples{
p <- ggplot(mtcars, aes(mpg, wt)) + geom_point()
save_col("mtcars-1col.png")  # single-column width
save_col("mtcars-1.5col.png", col_width = "2")  # double-column width
}
\keyword{ggplot2}
\keyword{ggsave}
\keyword{plot}
\keyword{save}
