% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Compute standard error of the mean}
\usage{
se(v)
}
\arguments{
\item{v}{Numerical vector to compute the standard error of.}
}
\description{
Easy function for computing standard error of the mean (how is this not in base R?).
}
\examples{
se(rnorm(n=50, mean=5, sd=1))
}
\keyword{error}
\keyword{standard}
