% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.R
\name{sem}
\alias{sem}
\title{Compute standard error of the mean}
\usage{
sem(v)
}
\arguments{
\item{v}{Numerical vector to compute the standard error of.}
}
\description{
Convenience function for computing standard error of the mean.
}
\examples{
sem(rnorm(n=50, mean=5, sd=1))
}
\keyword{error}
\keyword{standard}
