% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_theme.R
\name{simple_theme}
\alias{simple_theme}
\title{Custom ggplot2 theme}
\usage{
simple_theme()
}
\description{
Simple and clean customized ggplot2 theme.
}
\examples{
ggplot(mtcars, aes(wt, mpg)) + geom_point() + simple_theme()
}
\keyword{ggplot2}
\keyword{theme}
