% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pres.R
\name{theme_pres}
\alias{theme_pres}
\title{Custom ggplot2 theme for slide presentations}
\usage{
theme_pres()
}
\description{
ggplot2 theme with font size, linewidths, etc., optimized for slide
presentations.
}
\examples{
ggplot(mtcars, aes(wt, mpg)) + geom_point() + theme_pres()
}
\keyword{ggplot2}
\keyword{theme}
