% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pub.R
\name{theme_pub}
\alias{theme_pub}
\title{Custom ggplot2 theme for publication}
\usage{
theme_pub()
}
\description{
Publication-quality theme, modeled after the Nature final artwork guidelines.
See https://www.nature.com/nature/for-authors/final-submission for details.
}
\examples{
ggplot(mtcars, aes(wt, mpg)) + geom_point() + theme_pub()
}
\keyword{ggplot2}
\keyword{theme}
