% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pub_minimal.R
\name{theme_pub_minimal}
\alias{theme_pub_minimal}
\title{Custom ggplot2 theme for publication. Alternate of theme_pub() that is similar
to theme_minimal()}
\usage{
theme_pub_minimal()
}
\description{
Publication-quality theme, modeled after the Nature final artwork guidelines.
See https://www.nature.com/nature/for-authors/final-submission for details.
}
\examples{
ggplot(mtcars, aes(wt, mpg)) + geom_point() + theme_pub_minimal()
}
\keyword{ggplot2}
\keyword{theme}
