% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagesc.R
\name{imagesc}
\alias{imagesc}
\title{Plot matrix using ggplot2}
\usage{
imagesc(df)
}
\arguments{
\item{df}{Matrix data that has been converted to tidy dataframe using convert_mat_df}
}
\description{
Plot matrix analagous to Matlab's imagesc function.
}
\examples{
foo <- convert_mat_df(matrix(data=rnorm(n=25), nrow=5, ncol=5))
imagesc(foo)
}
\keyword{ggplot2}
\keyword{matrix}
