% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_chang.R
\name{theme_chang}
\alias{theme_chang}
\title{Chang theme for ggplot2}
\usage{
theme_chang(base_size = 11, base_family = "",
  base_line_size = base_size/22, base_rect_size = base_size/22)
}
\description{
Custom ggplot2 theme inspired by the plots in Raymond Chang's "Physical
Chemistry for the Chemical and Biological Sciences."
}
\details{
theme_chang() is a minimal design suited for illustrating single lines,
functions, or trends, or distinguishing one particular variable. Theme
includes the distinctive red used throughout the book, inward facing axis
ticks, and blank background.

theme_chang_minimal() is the same as theme_chang() but without axis ticks
and axis labels, useful for illustrative plots to show trends where exact
quantitation is not important.
}
\examples{
ggplot(mtcars, aes(wt, mpg)) + geom_point() + theme_chang()
ggplot(mtcars, aes(wt, mpg)) + geom_point() + theme_chang_minimal()
}
\keyword{ggplot2}
\keyword{theme}
